requirements.txt
openai
flask
flask-cors
python-dotenv
gunicorn
websockets
deepgram-sdk